refactor: rendererを分離
