+++ src/backend/electron/renderer/menuBarData.ts
